+++ R/helpers.R
@@ -147,9 +147,9 @@
 # remove NULL elements from lists
 .compact_list <- function(x, remove_na = FALSE) {
   if (remove_na) {
-    x[!sapply(x, function(i) length(i) == 0 || is.null(i) || (length(i) == 1 & is.na(i)) || any(i == "NULL"))]
+    x[!sapply(x, function(i) length(i) == 0 || is.null(i) || (length(i) == 1 & is.na(i)) || any(i == "NULL", na.rm = TRUE))]
   } else {
-    x[!sapply(x, function(i) length(i) == 0 || is.null(i) || any(i == "NULL"))]
+    x[!sapply(x, function(i) length(i) == 0 || is.null(i) || any(i == "NULL", na.rm = TRUE))]
   }
 }
 
@@ -169,4 +169,15 @@
     x$Parameter <- gsub("`", "", x$Parameter, fixed = TRUE)
   }
   x
+}
+
+
+#' @importFrom insight get_sigma
+.get_sigma <- function(model) {
+  s <- insight::get_sigma(model)
+  if (!is.null(s)) {
+    as.numeric(s)
+  } else {
+    NULL
+  }
 }