+++ R/model_performance.lm.R
@@ -3,7 +3,7 @@
 #' Compute indices of model performance for regression models.
 #'
 #' @param model A model.
-#' @param metrics Can be \code{"all"}, \code{"common"} or a character vector of metrics to be computed (some of \code{c("AIC", "BIC", "R2", "RMSE", "LOGLOSS", "PCP", "SCORE")}). \code{"common"} will compute AIC, BIC, R2 and RMSE.
+#' @param metrics Can be \code{"all"}, \code{"common"} or a character vector of metrics to be computed (some of \code{c("AIC", "BIC", "R2", "RMSE", "SIGMA", "LOGLOSS", "PCP", "SCORE")}). \code{"common"} will compute AIC, BIC, R2 and RMSE.
 #' @param verbose Toggle off warnings.
 #' @param ... Arguments passed to or from other methods.
 #'
@@ -16,6 +16,7 @@
 #'   \item{\strong{R2}} {r-squared value, see \code{\link{r2}}}
 #'   \item{\strong{R2_adj}} {adjusted r-squared, see \code{\link{r2}}}
 #'   \item{\strong{RMSE}} {root mean squared error, see \code{\link{performance_rmse}}}
+#'   \item{\strong{SIGMA}} {residual standard deviation, see \code{\link[insight:get_sigma]{get_sigma()}}}
 #'   \item{\strong{LOGLOSS}} {Log-loss, see \code{\link{performance_logloss}}}
 #'   \item{\strong{SCORE_LOG}} {score of logarithmic proper scoring rule, see \code{\link{performance_score}}}
 #'   \item{\strong{SCORE_SPHERICAL}} {score of spherical proper scoring rule, see \code{\link{performance_score}}}
@@ -31,8 +32,12 @@
 #' @importFrom insight model_info
 #' @export
 model_performance.lm <- function(model, metrics = "all", verbose = TRUE, ...) {
+  if (any(tolower(metrics) == "log_loss")) {
+    metrics[tolower(metrics) == "log_loss"] <- "LOGLOSS"
+  }
+
   if (all(metrics == "all")) {
-    metrics <- c("AIC", "BIC", "R2", "R2_adj", "RMSE", "LOGLOSS", "PCP", "SCORE")
+    metrics <- c("AIC", "BIC", "R2", "R2_adj", "RMSE", "SIGMA", "LOGLOSS", "PCP", "SCORE")
   } else if (all(metrics == "common")) {
     metrics <- c("AIC", "BIC", "R2", "R2_adj", "RMSE")
   }
@@ -61,16 +66,18 @@ model_performance.lm <- function(model, metrics = "all", verbose = TRUE, ...) {
   if ("RMSE" %in% toupper(metrics)) {
     out$RMSE <- performance_rmse(model, verbose = verbose)
   }
+  if ("SIGMA" %in% toupper(metrics)) {
+    out$Sigma <- .get_sigma(model)
+  }
   if (("LOGLOSS" %in% toupper(metrics)) && isTRUE(info$is_binomial)) {
     .logloss <- performance_logloss(model, verbose = verbose)
-    if (!is.na(.logloss)) out$LOGLOSS <- .logloss
+    if (!is.na(.logloss)) out$Log_loss <- .logloss
   }
   if (("SCORE" %in% toupper(metrics)) && (isTRUE(info$is_binomial) || isTRUE(info$is_count))) {
     .scoring_rules <- performance_score(model, verbose = verbose)
-    if (!is.na(.scoring_rules$logarithmic)) out$SCORE_LOG <- .scoring_rules$logarithmic
-    if (!is.na(.scoring_rules$spherical)) out$SCORE_SPHERICAL <- .scoring_rules$spherical
+    if (!is.na(.scoring_rules$logarithmic)) out$Score_log <- .scoring_rules$logarithmic
+    if (!is.na(.scoring_rules$spherical)) out$Score_spherical <- .scoring_rules$spherical
   }
-
   if (("PCP" %in% toupper(metrics)) && isTRUE(info$is_binomial) && !isTRUE(info$is_multinomial) && !isTRUE(info$is_ordinal)) {
     out$PCP <- performance_pcp(model, verbose = verbose)$pcp_model
   }
@@ -158,6 +165,14 @@ model_performance.flexsurvreg <- model_performance.lm
 
 #' @export
 model_performance.hurdle <- model_performance.lm
+
+#' @export
+model_performance.zeroinfl <- model_performance.lm
+
+#' @export
+model_performance.zerotrunc <- model_performance.lm
+
+