+++ R/model_performance.mixed.R
@@ -2,7 +2,7 @@
 #'
 #' Compute indices of model performance for mixed models.
 #'
-#' @param metrics Can be \code{"all"}, \code{"common"} or a character vector of metrics to be computed (some of \code{c("AIC", "BIC", "R2", "ICC", "RMSE", "LOGLOSS", "SCORE")}). \code{"common"} will compute AIC, BIC, R2, ICC and RMSE.
+#' @param metrics Can be \code{"all"}, \code{"common"} or a character vector of metrics to be computed (some of \code{c("AIC", "BIC", "R2", "ICC", "RMSE", "SIGMA", "LOGLOSS", "SCORE")}). \code{"common"} will compute AIC, BIC, R2, ICC and RMSE.
 #' @param ... Arguments passed to or from other methods.
 #' @inheritParams r2_nakagawa
 #' @inheritParams model_performance.lm
@@ -24,8 +24,12 @@
 #' @importFrom insight model_info
 #' @export
 model_performance.merMod <- function(model, metrics = "all", verbose = TRUE, ...) {
+  if (any(tolower(metrics) == "log_loss")) {
+    metrics[tolower(metrics) == "log_loss"] <- "LOGLOSS"
+  }
+
   if (all(metrics == "all")) {
-    metrics <- c("AIC", "BIC", "R2", "ICC", "RMSE", "LOGLOSS", "SCORE")
+    metrics <- c("AIC", "BIC", "R2", "ICC", "RMSE", "SIGMA", "LOGLOSS", "SCORE")
   } else if (all(metrics == "common")) {
     metrics <- c("AIC", "BIC", "R2", "ICC", "RMSE")
   }
@@ -50,13 +54,16 @@ model_performance.merMod <- function(model, metrics = "all", verbose = TRUE, ...
   if ("RMSE" %in% metrics) {
     out$RMSE <- performance_rmse(model, verbose = verbose)
   }
+  if ("SIGMA" %in% toupper(metrics)) {
+    out$Sigma <- .get_sigma(model)
+  }
   if (("LOGLOSS" %in% metrics) && mi$is_binomial) {
-    out$LOGLOSS <- performance_logloss(model, verbose = verbose)
+    out$Log_loss <- performance_logloss(model, verbose = verbose)
   }
   if (("SCORE" %in% metrics) && (mi$is_binomial || mi$is_count)) {
     .scoring_rules <- performance_score(model, verbose = verbose)
-    if (!is.na(.scoring_rules$logarithmic)) out$SCORE_LOG <- .scoring_rules$logarithmic
-    if (!is.na(.scoring_rules$spherical)) out$SCORE_SPHERICAL <- .scoring_rules$spherical
+    if (!is.na(.scoring_rules$logarithmic)) out$Score_log <- .scoring_rules$logarithmic
+    if (!is.na(.scoring_rules$spherical)) out$Score_spherical <- .scoring_rules$spherical
   }
 
   # TODO: What with sigma and deviance?
@@ -91,6 +98,10 @@ model_performance.glmmTMB <- model_performance.merMod
 
 #' @export
 model_performance.mixor <- function(model, metrics = "all", verbose = TRUE, ...) {
+  if (any(tolower(metrics) == "log_loss")) {
+    metrics[tolower(metrics) == "log_loss"] <- "LOGLOSS"
+  }
+
   if (all(metrics == "all")) {
     metrics <- c("AIC", "BIC", "LOGLOSS", "SCORE")
   }
@@ -105,12 +116,12 @@ model_performance.mixor <- function(model, metrics = "all", verbose = TRUE, ...)
     out$BIC <- .get_BIC(model)
   }
   if (("LOGLOSS" %in% metrics) && mi$is_binomial && !mi$is_ordinal && !mi$is_multinomial) {
-    out$LOGLOSS <- performance_logloss(model, verbose = verbose)
+    out$Log_loss <- performance_logloss(model, verbose = verbose)
   }
   if (("SCORE" %in% metrics) && (mi$is_binomial || mi$is_count) && !mi$is_ordinal && !mi$is_multinomial) {
     .scoring_rules <- performance_score(model, verbose = verbose)
-    if (!is.na(.scoring_rules$logarithmic)) out$SCORE_LOG <- .scoring_rules$logarithmic
-    if (!is.na(.scoring_rules$spherical)) out$SCORE_SPHERICAL <- .scoring_rules$spherical
+    if (!is.na(.scoring_rules$logarithmic)) out$Score_log <- .scoring_rules$logarithmic
+    if (!is.na(.scoring_rules$spherical)) out$Score_spherical <- .scoring_rules$spherical
   }
 
   out <- as.data.frame(out)