+++ R/r2.R
@@ -383,6 +383,9 @@ r2.brmsfit <- function(model, ...) {
 r2.stanreg <- r2.brmsfit
 
 
+#' @export
+r2.BFBayesFactor <- r2.brmsfit
+
 
 
 
@@ -402,6 +405,9 @@ r2.gam <- function(model, ...) {
     NextMethod()
   }
 }
+
+#' @export
+r2.scam <- r2.gam
 
 
 #' @export