+++ R/r2_bayes.R
@@ -3,11 +3,15 @@
 #'
 #' @description Compute R2 for Bayesian models. For mixed models (including a random part),
 #' it additionally computes the R2 related to the fixed effects only (marginal R2).
+#' While \code{r2_bayes()} returns a single R2 value, \code{r2_posterior()} returns
+#' a posterior sample of Bayesian R2 values.
 #'
-#' @param model A Bayesian regression model.
+#' @param model A Bayesian regression model (from \code{brms}, \code{rstanarm},
+#'   \code{BayesFactor}, etc).
 #' @param robust Logical, if \code{TRUE}, the median instead of mean is used to
 #'   calculate the central tendency of the variances.
 #' @param ci Value or vector of probability of the CI (between 0 and 1) to be estimated.
+#' @param ... Arguments passed to \code{r2_posterior()}.
 #'
 #' @return A list with the Bayesian R2 value. For mixed models, a list with the
 #'   Bayesian R2 value and the marginal Bayesian R2 value. The standard errors
@@ -20,6 +24,9 @@
 #'   For mixed models, the conditional and marginal R2 are returned. The marginal
 #'   R2 considers only the variance of the fixed effects, while the conditional
 #'   R2 takes both the fixed and random effects into account.
+#'   \cr \cr
+#'   \code{r2_posterior()} is the actual workhorse for \code{r2_bayes()} and
+#'   returns a posterior sample of Bayesian R2 values.
 #'
 #' @examples
 #' library(performance)
@@ -36,7 +43,31 @@
 #'   )
 #'   r2_bayes(model)
 #' }
+#'
+#'
 #' \dontrun{
+#' if (require("BayesFactor")) {
+#'   data(mtcars)
+#'
+#'   BFM <- generalTestBF(mpg ~ qsec + gear, data = mtcars, progress = FALSE)
+#'   FM <- lm(mpg ~ qsec + gear, data = mtcars)
+#'
+#'   r2_bayes(FM)
+#'   r2_bayes(BFM[3])
+#'   r2_bayes(BFM, average = TRUE) # across all models
+#'
+#'
+#'   # with random effects:
+#'   mtcars$gear <- factor(mtcars$gear)
+#'   model <- lmBF(
+#'     mpg ~ hp + cyl + gear + gear:wt,
+#'     mtcars,
+#'     progress = FALSE,
+#'     whichRandom = c("gear", "gear:wt")
+#'   )
+#'   r2_bayes(model)
+#' }
+#'
 #' if (require("brms")) {
 #'   model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
 #'   r2_bayes(model)
@@ -47,12 +78,12 @@
 #' }
 #' @references Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2018). R-squared for Bayesian regression models. The American Statistician, 1–6. \doi{10.1080/00031305.2018.1549100}
 #'
-#' @importFrom insight find_algorithm is_multivariate find_response
+#' @importFrom insight find_algorithm is_multivariate find_response model_info get_response find_predictors
 #' @importFrom stats median mad sd
-#' @importFrom bayestestR ci
+#' @importFrom bayestestR ci hdi point_estimate
 #' @export
-r2_bayes <- function(model, robust = TRUE, ci = .89) {
-  r2_bayesian <- .r2_posterior(model)
+r2_bayes <- function(model, robust = TRUE, ci = .89, ...) {
+  r2_bayesian <- r2_posterior(model, ...)
 
   if (is.null(r2_bayesian)) {
     return(NULL)
@@ -63,7 +94,7 @@ r2_bayes <- function(model, robust = TRUE, ci = .89) {
       class = "r2_bayes_mv",
       rapply(r2_bayesian, ifelse(robust, stats::median, mean)),
       "SE" = rapply(r2_bayesian, ifelse(robust, stats::mad, stats::sd)),
-      "Estimates" = rapply(r2_bayesian, bayestestR::point_estimate, centrality = "all", dispersion = TRUE),
+      # "Estimates" = rapply(r2_bayesian, bayestestR::point_estimate, centrality = "all", dispersion = TRUE),
       "CI" = rapply(r2_bayesian, bayestestR::hdi, ci = ci),
       "robust" = robust
     )
@@ -72,16 +103,22 @@ r2_bayes <- function(model, robust = TRUE, ci = .89) {
       class = "r2_bayes",
       lapply(r2_bayesian, ifelse(robust, stats::median, mean)),
       "SE" = lapply(r2_bayesian, ifelse(robust, stats::mad, stats::sd)),
-      "Estimates" = lapply(r2_bayesian, bayestestR::point_estimate, centrality = "all", dispersion = TRUE),
+      # "Estimates" = lapply(r2_bayesian, bayestestR::point_estimate, centrality = "all", dispersion = TRUE),
       "CI" = lapply(r2_bayesian, bayestestR::hdi, ci = ci),
       "robust" = robust
     )
   }
 }
 
+#' @export
+#' @rdname r2_bayes
+r2_posterior <- function(model, ...){
+  UseMethod("r2_posterior")
+}
 
-
-.r2_posterior <- function(model) {
+#' @export
+#' @rdname r2_bayes
+r2_posterior.brmsfit <- function(model, ...) {
   if (!requireNamespace("rstantools", quietly = TRUE)) {
     stop("Package `rstantools` needed for this function to work. Please install it.")
   }
@@ -141,4 +178,142 @@ r2_bayes <- function(model, robust = TRUE, ci = .89) {
       NULL
     }
   )
+}
+
+
+#' @export
+#' @rdname r2_bayes
+r2_posterior.stanreg <- r2_posterior.brmsfit
+
+#' @param average Compute model-averaged index? See
+#'   \code{\link[bayestestR:weighted_posteriors]{bayestestR::weighted_posteriors()}}.
+#' @inheritParams bayestestR::weighted_posteriors
+#' @importFrom insight get_parameters get_response find_predictors
+#' @importFrom stats median mad sd
+#' @importFrom bayestestR point_estimate hdi
+#' @importFrom utils packageVersion
+#' @export
+#' @rdname r2_bayes
+r2_posterior.BFBayesFactor <- function(model, average = FALSE, prior_odds = NULL, ...){
+  if (average) {
+    return(.r2_posterior_model_average(model, prior_odds = prior_odds))
+  }
+
+  if (!requireNamespace("rstantools", quietly = TRUE)) {
+    stop("Package `rstantools` needed for this function to work. Please install it.")
+  }
+
+  if (!requireNamespace("BayesFactor", quietly = TRUE)) {
+    stop("Package `BayesFactor` needed for this function to work. Please install it.")
+  }
+
+  # Estimates
+  params <- insight::get_parameters(model, unreduce = FALSE)
+  # remove sig and g cols
+  params <- params[, !grepl(pattern = "^sig2$|^g_|^g$", colnames(params))]
+
+  # Model Matrix
+  mm <- BayesFactor::model.matrix(model[1])
+  colnames(mm)[1] <- "mu"
+
+  # match?
+  if ((length(colnames(params)) != length(colnames(mm))) ||
+      !all(colnames(params) == colnames(mm))) {
+    if (utils::packageVersion("BayesFactor") < package_version("0.9.12.4.3")) {
+      stop("R2 for BayesFactor models with random effects requires BayesFactor v0.9.12.4.3 or higher.", call. = FALSE)
+    }
+    stop("Woops, you seem to have stumbled on some weird edge case. Please file an issue at https://github.com/easystats/performance/issues", call. = FALSE)
+  }
+
+  # Compute R2!
+  y <- insight::get_response(model)
+  yy <- as.matrix(params) %*% t(mm)
+  r2s <- rstantools::bayes_R2(yy, y = y)
+  r2_bayesian <- data.frame(R2_Bayes = r2s)
+
+  rand <- insight::find_predictors(model[1], effects = "random", flatten = TRUE)
+  if (!is.null(rand)) {
+    idx <- sapply(paste0("\\b", rand, "\\b"), grepl, x = colnames(params))
+    idx <- apply(idx, 1, any)
+    params[idx] <- 0
+
+    yy <- as.matrix(params) %*% t(mm)
+    r2s_marginal <- rstantools::bayes_R2(yy, y = y)
+    r2_bayesian$R2_Bayes_marginal <- r2s_marginal
+  }
+
+  r2_bayesian
+}
+
+
+
+#' @importFrom bayestestR weighted_posteriors
+#' @keywords internal
+.r2_posterior_model_average <- function(model, prior_odds = NULL) {
+  if (!requireNamespace("BayesFactor", quietly = TRUE)) {
+    stop("Package `BayesFactor` needed for this function to work. Please install it.")
+  }
+
+  BFMods <- bayestestR::bayesfactor_models(model, verbose = FALSE)
+  has_random <- !is.null(insight::find_predictors(model, effects = "random", flatten = TRUE))
+
+  # extract parameters
+  intercept_only <- which(BFMods$Model == "1")
+  params <- vector(mode = "list", length = nrow(BFMods))
+  for (m in seq_along(params)) {
+    if (length(intercept_only) && m == intercept_only) {
+      params[[m]] <- data.frame(R2_Bayes = rep(0, 4000))
+    } else if (m == 1) {
+      # If the model is the "den" model
+      params[[m]] <- suppressMessages(r2_posterior(1 / model[1]))
+    } else {
+      params[[m]] <- suppressMessages(r2_posterior(model[m - 1]))
+    }
+
+    # when there is no random effect, marginal = conditional
+    if (has_random && is.null(params[[m]]$R2_Bayes_marginal)) {
+      params[[m]]$R2_Bayes_marginal <- params[[m]]$R2_Bayes
+    }
+  }
+
+
+  # Compute posterior model probabilities
+  if (!is.null(prior_odds)) {
+    prior_odds <- c(1, prior_odds)
+  } else {
+    prior_odds <- rep(1, nrow(BFMods))
+  }
+  posterior_odds <- prior_odds * BFMods$BF
+  posterior_odds <- posterior_odds[-1] / posterior_odds[1]
+
+  do.call(bayestestR::weighted_posteriors,
+          c(params, list(missing = 0, prior_odds = posterior_odds)))
+}
+
+
+
+#' @export
+as.data.frame.r2_bayes <- function(x, ...) {
+  out <- data.frame(
+    R2 = x$R2_Bayes,
+    SD = attributes(x)$SE$R2_Bayes,
+    CI = attributes(x)$CI$R2_Bayes$CI,
+    CI_low = attributes(x)$CI$R2_Bayes$CI_low,
+    CI_high = attributes(x)$CI$R2_Bayes$CI_high,
+    stringsAsFactors = FALSE
+  )
+  if (!is.null(x$R2_Bayes_marginal)) {
+    out_marginal <- data.frame(
+      R2 = x$R2_Bayes_marginal,
+      SD = attributes(x)$SE$R2_Bayes_marginal,
+      CI = attributes(x)$CI$R2_Bayes_marginal$CI,
+      CI_low = attributes(x)$CI$R2_Bayes_marginal$CI_low,
+      CI_high = attributes(x)$CI$R2_Bayes_marginal$CI_high,
+      stringsAsFactors = FALSE
+    )
+    out$Component <- "conditional"
+    out_marginal$Component <- "marginal"
+    out <- rbind(out, out_marginal)
+  }
+  out
 }